migrate test_xapi_virt to pytest
